backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
mysqlclient>=2.2.0
PyJWT>=2.8.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
waitress>=3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pandas>=2.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pywebpush>=1.14.0

